Rough in File\Target
Create an adapter to make a target work as a dependency for another
build. This isn't complete and has some serious gaps around caching and
performance as the intermediate build is rebuilt each time the desired
target is compiled. Addressing this problem will require more extensive
rework in the Compiler as it is not aware of caching/output directories
right now.