Start upgrading to php8.1 types
I'm thinking that long term, I'll want a version that is php8.1+.
This will likely require a major version bump, but I could maintain 1.x
as <php8.1.

Update to cakephp codesniffer as it has done almost all of this work
using fixers. There is still a lot of missing types, and then begins the
process of aligning types and what the tests are expecting.